+++ .idea/gradle.xml
@@ -1,18 +1,19 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
-	<component name="GradleSettings">
-		<option name="linkedExternalProjectsSettings">
-			<GradleProjectSettings>
-				<option name="testRunner" value="GRADLE" />
-				<option name="distributionType" value="DEFAULT_WRAPPED" />
-				<option name="externalProjectPath" value="$PROJECT_DIR$" />
-				<option name="modules">
-					<set>
-						<option value="$PROJECT_DIR$" />
-						<option value="$PROJECT_DIR$/app" />
-					</set>
-				</option>
-			</GradleProjectSettings>
-		</option>
-	</component>
+  <component name="GradleMigrationSettings" migrationVersion="1" />
+  <component name="GradleSettings">
+    <option name="linkedExternalProjectsSettings">
+      <GradleProjectSettings>
+        <option name="testRunner" value="GRADLE" />
+        <option name="distributionType" value="DEFAULT_WRAPPED" />
+        <option name="externalProjectPath" value="$PROJECT_DIR$" />
+        <option name="modules">
+          <set>
+            <option value="$PROJECT_DIR$" />
+            <option value="$PROJECT_DIR$/app" />
+          </set>
+        </option>
+      </GradleProjectSettings>
+    </option>
+  </component>
 </project>
