+++ app/src/androidTest/java/com/spiral_root/android/test_app/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.spiral_root.android.slf4j_timber
+package com.spiral_root.android.test_app
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
