+++ app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
 	xmlns:tools="http://schemas.android.com/tools"
-	package="com.spiral_root.android.slf4j_timber">
+	package="com.spiral_root.android.test_app">
 
 	<application
 		android:allowBackup="true"
@@ -11,13 +11,13 @@
 		android:label="@string/app_name"
 		android:roundIcon="@mipmap/ic_launcher_round"
 		android:supportsRtl="true"
-		android:theme="@style/Theme.Slf4jtimber"
+		android:theme="@style/Theme.TestApp"
 		tools:targetApi="31">
 		<activity
-			android:name=".MainActivity"
+			android:name="com.spiral_root.android.test_app.MainActivity"
 			android:exported="true"
 			android:label="@string/app_name"
-			android:theme="@style/Theme.Slf4jtimber">
+			android:theme="@style/Theme.TestApp">
 			<intent-filter>
 				<action android:name="android.intent.action.MAIN" />
