+++ app/src/main/java/com/spiral_root/android/test_app/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.spiral_root.android.slf4j_timber.ui.theme
+package com.spiral_root.android.test_app.ui.theme
 
 import androidx.compose.ui.graphics.Color
