+++ app/src/main/java/com/spiral_root/android/test_app/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.spiral_root.android.slf4j_timber.ui.theme
+package com.spiral_root.android.test_app.ui.theme
 
 import android.app.Activity
 import android.os.Build
@@ -38,7 +38,7 @@ private val LightColorScheme = lightColorScheme(
 )
 
 @Composable
-fun Slf4jtimberTheme(
+fun TestAppTheme(
 	darkTheme: Boolean = isSystemInDarkTheme(),
 	// Dynamic color is available on Android 12+
 	dynamicColor: Boolean = true,
